Rework our TranslationsModel and RecitationsModel to get rid of the hacks
We previously duplicated all the items. The first half of the model was representing
the installed/enabled ones and the second half representing the rest.

This was a hack and has been fixed now.

We now do not duplicate anything and each model can do its sorting
using a newly introduced interface (LessThanInterface).
Also rename VisibilityFilterModel to SortModel.
SortModel takes care of the sorting now using the LessThanInterface implemented
by the source model.