Rework Alsa::hasFrames()
We depen on the fact that an underrun will occur and we will detect that. This happens only
if we have fed frames to alsa.
If we -however- start playback but no frames have been fed because the decoder failed
to decode the first frame then we will never get an underrun.
Therefor we need to compare the available frames to the buffer size