Get rid of mirrored layout
We had to use a mirrored layout because we need RTL navigation direction.
It was however causing a lot of issues, most notably the page not being updated
when we navigate to a bookmark or even a chapter from the index page.
An animation (like the toolbar expand one) was however enough to cause a redraw.

Let's get rid of the mirrored layout and do some clever calculations to get the
same needed RTL navigation