MediaPlayer: Fix pausedChanged() emission when we stop playback
We emit the signal before destroying the audio output object
If we are paused then the signal emission will trigger reading the paused
property which will still return true.
What we need to do is to emit the signal after destroying the audio output object.
We emit it only if we were already paused.