android: rework binder validity check
It is better to protect the sender binder with a lock since
ServiceConnection is multithreaded.

We are also working around a crash by not using QAndroidBinder handle().isValid() and
using a boolean to track the validity of the binder instead.

Symptoms:
Remote service crash will cause an app crash somewhere inside onTransact()
It seems that isValid() is not returning false even though it should which
causes us to try to use the binder object leading to a crash.