downloader: rework download progress calculation
If bytes total is 0 then we end up assigning it to m_size and happily
dividing bytes received by that ending up in a division by zero crash.
his has been fixed now.

While at it, get rid of Translation::downloadSize as it is not used.

Also neither try to figure out the download total if Qt does not report it
nor should we hard code a size. Simply do not emit any signal.