Implement flip to pause playback
We leverage the existing PhoneFlipControl after refactoring and renaming
it to FlipSensor Each MediaService needs to implement the handling. The
desktop service is straight forward.

As for Android, FlipSensor needs to be handled by the actual service.
The service needed a new UpdateSettings command which is transmitted
to the actual service along with a Bundle that carries the flip to
pause setting only for now but could be expanded easily in the future.